backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.13.0

# AI/ML Libraries
openai>=1.12.0
anthropic>=0.72.0
langchain>=0.1.0
pypdf2==3.0.1
pdfplumber==0.10.3
google-genai>=1.0.0
chromadb>=0.4.0

# API Integrations
requests==2.31.0
httpx>=0.27.0

# Background Jobs
celery==5.3.4
redis==5.0.1

# Report Generation
reportlab==4.0.7
openpyxl==3.1.2
pandas>=2.0.0
python-docx>=1.1.0

# Utilities
python-dateutil==2.8.2
email-validator>=2.0.0
